# inets: Speed up test suite by ~ 2 minutes