--- a/lib/inets/test/httpd_SUITE.erl
+++ b/lib/inets/test/httpd_SUITE.erl
@@ -79,11 +79,8 @@ all() ->
      {group, http_not_sup},
      {group, https_alert},
      {group, https_not_sup},
-     mime_types_format,
-     erl_script_timeout_default,
-     erl_script_timeout_option,
-     erl_script_timeout_proplist,
-     erl_script_alias_all
+     {group, esi},
+     mime_types_format
     ].
 
 groups() ->
@@ -111,7 +108,7 @@ groups() ->
      {http_not_sup, [], [{group, not_sup}]},
      {https_not_sup, [], [{group, not_sup}]},
      {https_alert, [], [tls_alert]},
-     {http_mime_types, [], [alias_1_1, alias_1_0]},
+     {http_mime_types, [parallel], [alias_1_1, alias_1_0]},
      {limit, [],  [content_length, max_clients_1_1]},
      {custom, [],  [customize, add_default]},
      {reload, [], [non_disturbing_reconfiger_dies,
@@ -137,6 +134,10 @@ groups() ->
       ++ http_head() ++ http_get() ++ load()},
      {http_1_0, [], [host, cgi, trace] ++ http_head() ++ http_get() ++ load()},
      {http_rel_path_script_alias, [], [cgi]},
+     {esi, [], [erl_script_timeout_default,
+                erl_script_timeout_option,
+                erl_script_timeout_proplist,
+                erl_script_alias_all]},
      {not_sup, [], [put_not_sup]}
     ].
 
@@ -262,6 +263,9 @@ init_per_group(http_rel_path_script_alias = Group, Config) ->
     init_httpd(Group, [{type, ip_comm},{http_version, "HTTP/1.1"}| Config]);
 init_per_group(not_sup, Config) ->
     [{http_version, "HTTP/1.1"} | Config];
+init_per_group(Group, Config) when Group == esi ->
+    ok = start_apps(Group),
+    Config;
 init_per_group(_, Config) ->
     Config.
 
@@ -274,7 +278,8 @@ end_per_group(Group, _Config)  when  Group == http_basic;
 				     Group == http_security;
 				     Group == http_reload;
                                      Group == http_post;
-                                     Group == http_mime_types
+                                     Group == http_mime_types;
+                                     Group == esi
 				     ->
     inets:stop();
 end_per_group(Group, _Config) when  Group == https_basic;
@@ -1791,20 +1796,14 @@ mime_types_format(Config) when is_list(Config) ->
      {"hqx","application/mac-binhex40"}]} = httpd_conf:load_mime_types(MimeTypes).
 
 erl_script_timeout_default(Config) when is_list(Config) ->
-    inets:start(),
-    {ok, Pid} =	inets:start(httpd,
-                            [{port, 0},
-                             {server_name,"localhost"},
-                             {server_root,"./"},
-                             {document_root,"./"},
-                             {bind_address, any},
-                             {mimetypes, [{"html", "text/html"}]},
-                             {modules,[mod_esi]},
-                             {erl_script_alias, {"/erl", [httpd_example]}}
-                            ]),
-    Info = httpd:info(Pid),
+    ServerConfig = [
+        {modules, [mod_esi]},
+        {erl_script_alias, {"/erl", [httpd_example]}}
+        | Config
+    ],
+    Httpd = init_httpd(esi, ServerConfig),
 
-    Port = proplists:get_value(port, Info),
+    Port = proplists:get_value(port, Httpd),
 
     %% Default erl_script_timeout is 15.
     %% Verify:  13 =< erl_script_timeout =< 17
@@ -1812,22 +1811,17 @@ erl_script_timeout_default(Config) when is_list(Config) ->
 
     {ok, {_, _, Body}} = httpc:request(get, {Url, []}, [{timeout, 45000}], []),
     ct:log("Response: ~p~n", [Body]),
-    verify_body(Body, 13000),
-    inets:stop().
+    verify_body(Body, 13000).
 
 erl_script_timeout_option(Config) when is_list(Config) ->
-    inets:start(),
-    {ok, Pid} =	inets:start(httpd,
-                            [{port, 0},
-                             {server_name,"localhost"},
-                             {server_root,"./"},
-                             {document_root,"./"},
-                             {bind_address, any},
-                             {mimetypes, [{"html", "text/html"}]},
-                             {modules,[mod_esi]},
-                             {erl_script_timeout, 2},
-                             {erl_script_alias, {"/erl", [httpd_example]}}
-                            ]),
+    ServerConfig = [
+        {modules, [mod_esi]},
+        {erl_script_timeout, 2},
+        {erl_script_alias, {"/erl", [httpd_example]}}
+        | Config
+    ],
+    ServerInfo = init_httpd(esi, ServerConfig),
+    Pid = proplists:get_value(server_pid, ServerInfo),
     Info = httpd:info(Pid),
     verify_timeout(Info, 2),
 
@@ -1838,8 +1832,7 @@ erl_script_timeout_option(Config) when is_list(Config) ->
 
     {ok, {_, _, Body}} = httpc:request(Url),
     ct:log("Response: ~p~n", [Body]),
-    verify_body(Body, 1000),
-    inets:stop().
+    verify_body(Body, 1000).
 
 erl_script_timeout_proplist(Config) when is_list(Config) ->
     HttpdConf = filename:join(get_tmp_dir(Config),
@@ -1857,7 +1850,6 @@ erl_script_timeout_proplist(Config) when is_list(Config) ->
         "].",
     ok = file:write_file(HttpdConf, ServerConfig),
 
-    inets:start(),
     {ok, Pid} =	inets:start(httpd,
                             [{proplist_file, HttpdConf}]),
     Info = httpd:info(Pid),
@@ -1870,8 +1862,7 @@ erl_script_timeout_proplist(Config) when is_list(Config) ->
 
     {ok, {_, _, Body}} = httpc:request(Url),
     ct:log("Response: ~p~n", [Body]),
-    verify_body(Body, 3000),
-    inets:stop().
+    verify_body(Body, 3000).
 
 erl_script_alias_all(Config0) when is_list(Config0) ->
     ok = start_apps(http_basic),
@@ -1880,8 +1871,7 @@ erl_script_alias_all(Config0) when is_list(Config0) ->
                Config0],
     Config2 = init_httpd(http_basic_erl_script_alias_all, Config1),
     ok = http_status("GET /cgi-bin/erl/httpd_example:get ",
-        	     Config2, [{statuscode, 200}]),
-    inets:stop().
+        	     Config2, [{statuscode, 200}]).
 
 tls_alert(Config) when is_list(Config) ->
     SSLOpts = proplists:get_value(client_alert_conf, Config),    
@@ -2036,7 +2026,8 @@ start_apps(Group) when  Group == http_basic;
                         Group == http_mime_types;
                         Group == http_rel_path_script_alias;
                         Group == http_not_sup;
-                        Group == http_mime_types->
+                        Group == http_mime_types;
+                        Group == esi ->
     inets_test_lib:start_apps([inets]).
 
 server_start(_, HttpdConfig) ->
@@ -2152,7 +2143,9 @@ server_config(https, Config) ->
     ServerConf = proplists:get_value(server_config, SSLConf),
     [{socket_type, {ssl,
 		    [{nodelay, true} | ServerConf]}}]
-        ++ proplists:delete(socket_type, server_config(http, Config)).
+        ++ proplists:delete(socket_type, server_config(http, Config));
+server_config(esi, Config) ->
+    basic_conf() ++ server_config(http, Config).
 
 config_template(Config, ServerRoot, ScriptPath, Modules) ->
     [{port, 0},
@@ -2173,7 +2166,15 @@ config_template(Config, ServerRoot, ScriptPath, Modules) ->
      {script_alias, {"/cgi-bin/", ScriptPath}},
      {script_re_write, {"/cgi-([a-zA-Z-]*)bin/", ScriptPath}},
      {erl_script_alias, {"/cgi-bin/erl", Modules}}
-    ].
+    ] ++ custom_config_options(Config).
+
+custom_config_options([{Name, _} = Option | Rest]) when Name == erl_script_alias;
+                                                        Name == erl_script_timeout ->
+    [Option | custom_config_options(Rest)];
+custom_config_options([_ | Rest]) ->
+    custom_config_options(Rest);
+custom_config_options([]) ->
+    [].
 
 init_httpd(Group, Config0) ->
     Config1 = proplists:delete(port, Config0),