--- a/lib/inets/test/httpd_basic_SUITE.erl
+++ b/lib/inets/test/httpd_basic_SUITE.erl
@@ -34,21 +34,20 @@ suite() -> [{ct_hooks,[ts_install_cth]},
 	    {timetrap, {seconds, 30}}].
 
 all() -> 
-    [uri_too_long_414, 
-     header_too_long_413,
-     entity_too_long,
-     http_0_9_not_supported,
-     erl_script_nocache_opt,
-     script_nocache,
-     escaped_url_in_error_body,
-     script_timeout,
-     slowdose,
-     keep_alive_timeout,
-     invalid_rfc1123_date
-    ].
+    [{group, httpd_basic}].
 
 groups() -> 
-    [].
+    [{httpd_basic, [parallel], [uri_too_long_414,
+                                header_too_long_413,
+                                entity_too_long,
+                                http_0_9_not_supported,
+                                erl_script_nocache_opt,
+                                script_nocache,
+                                escaped_url_in_error_body,
+                                script_timeout,
+                                slowdose,
+                                keep_alive_timeout,
+                                invalid_rfc1123_date]}].
 
 init_per_group(_GroupName, Config) ->
     Config.