--- a/lib/inets/test/inets_SUITE.erl
+++ b/lib/inets/test/inets_SUITE.erl
@@ -84,10 +84,6 @@ end_per_suite(_Config) ->
 %% Note: This function is free to add any key/value pairs to the Config
 %% variable, but should NOT alter/remove any existing entries.
 %%--------------------------------------------------------------------
-init_per_testcase(httpd_reload, Config) ->
-    inets:stop(),
-    ct:timetrap({seconds, 40}),
-    Config;
 init_per_testcase(_Case, Config) ->
     inets:stop(),
     Config.