Improve inets test suite runtime from 560s to 435s on my fast local machine, from 595s to 486s on my slow local machine, and from 10m47s to 8m49s on GitHub actions in my fork.
 
The following was done:
- Some testcases that supported it were just migrated to `parallel` directly
- Global state in tests is either isolated out of parallel groups or only taken when needed (ETS tables, httpc profiles)
- Removed or reduced naptime in some tests depending on shared resources
- Removed debug tracing from `esi_put`, this seemed to hang the tests when running in GH Actions

I've tried to keep the individual changes cleanly separated, let me know if they should be squashed.